Measurement conditions

+++ FiresWinForms/DataModel.cs
@@ -2,7 +2,7 @@
 {
     public class DataModel
     {
-        public double? Time { get; set; }
+        public decimal? Time { get; set; }
         public decimal? Value { get; set; }
     }
 }
